Rework BitmapFonts for new project organization.
