+++ source/MonoGame.Extended.Content.Pipeline/BitmapFonts/BitmapFontWriter.cs
@@ -1,6 +1,7 @@
 using Microsoft.Xna.Framework.Content.Pipeline;
 using Microsoft.Xna.Framework.Content.Pipeline.Serialization.Compiler;
 using MonoGame.Extended.BitmapFonts;
+using MonoGame.Extended.Content.ContentReaders;
 
 namespace MonoGame.Extended.Content.Pipeline.BitmapFonts
 {
@@ -49,7 +50,7 @@ namespace MonoGame.Extended.Content.Pipeline.BitmapFonts
 
         public override string GetRuntimeReader(TargetPlatform targetPlatform)
         {
-            return typeof(BitmapFontReader).AssemblyQualifiedName;
+            return typeof(BitmapFontContentReader).AssemblyQualifiedName;
         }
     }
 }
