--- source/MonoGame.Extended/BitmapFonts/BmfTypes/BmfKerningPairsBlock.cs
@@ -1,17 +1,0 @@
-﻿// Copyright (c) Craftwork Games. All rights reserved.
-// Licensed under the MIT license.
-// See LICENSE file in the project root for full license information.
-
-using System.Runtime.InteropServices;
-
-namespace MonoGame.Extended.BitmapFonts.BmfTypes;
-
-[StructLayout(LayoutKind.Explicit)]
-public struct BmfKerningPairsBlock
-{
-    public const int StructSize = 10;
-
-    [FieldOffset(0)] public uint First;
-    [FieldOffset(4)] public uint Second;
-    [FieldOffset(8)] public short Amount;
-}
